Refactor admin tutorials page into modular components
